Rename atomicops to nrtdynmod
